Yep its broke for Big Integer paramters.. saving todays work

+++ SymptomManagement/src/main/java/com/skywomantech/cloud/symptommanagement/repository/MedicationRepository.java
@@ -1,4 +1,5 @@
 package com.skywomantech.cloud.symptommanagement.repository;
+import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -6,10 +7,11 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 
+
 import com.skywomantech.cloud.symptommanagement.client.SymptomManagementApi;
 
 @RepositoryRestResource(path = SymptomManagementApi.MEDICATION_PATH)
-public interface MedicationRepository extends MongoRepository<Medication, Long>{
+public interface MedicationRepository extends MongoRepository<Medication, BigInteger>{
 
 		// Find all medications with a similar name
 		public Collection<Medication> findByName(
